:sparkles: Create metadata
and change generateStaticParams
