+++ content/cheatsheet/zod.mdx
@@ -8,3 +8,5 @@ tags: react
 # zod
 
 zod는 유효성 검증을 도와주는 라이브러리입니다.
+
+엄준식
