+++ content/post/html-video.mdx
@@ -119,8 +119,6 @@ width에 vw를 100을 주게되면 가로로 꽉찬 화면을 보게되죠!
 
 object-fit 속성은 아주 강력한 속성입니다.
 
-이미지를 콘텐츠
-
 object-fit 속성을 cover로 부여하면 콘텐츠의 가로세로비는 유지하면서 콘텐츠 박스를 가득채우게 됩니다.
 
 만약 가로세로비가 일치하지 않으면 요소를 조금 확대하고 가장자리를 자릅니다.
@@ -134,3 +132,5 @@ https://developer.mozilla.org/ko/docs/Web/CSS/object-fit
 즉 부모속성의 크기가 정해져있는 경우에는 화면을 꽉채우지 않을 수 있습니다.
 
 이 점 유의하시기 바랍니다!
+
+## 마치며
